+++ src/Ouzo/Core/Db/StatementIterator.php
@@ -1,0 +1,46 @@
+<?php
+/*
+ * Copyright (c) Ouzo contributors, http://ouzoframework.org
+ * This file is made available under the MIT License (view the LICENSE file for more information).
+ */
+namespace Ouzo\Db;
+
+use Iterator;
+use IteratorIterator;
+use PDOStatement;
+
+class StatementIterator implements Iterator
+{
+
+    private $iterator;
+
+    public function __construct(PDOStatement $statement)
+    {
+        $this->iterator = new IteratorIterator($statement);
+    }
+
+    public function current()
+    {
+        return $this->iterator->current();
+    }
+
+    public function next()
+    {
+        $this->iterator->next();
+    }
+
+    public function key()
+    {
+        return $this->iterator->key();
+    }
+
+    public function valid()
+    {
+        return $this->iterator->valid();
+    }
+
+    public function rewind()
+    {
+        $this->iterator->rewind();
+    }
+}